+++ gherkin-utils/javascript/src/Query.ts
@@ -2,10 +2,11 @@ import { messages } from '@cucumber/messages'
 import { ArrayMultimap } from '@teppeis/multimaps'
 
 export default class Query {
+  private readonly sources: messages.ISource[] = []
   private readonly gherkinDocuments: messages.IGherkinDocument[] = []
   private readonly pickles: messages.IPickle[] = []
   private readonly locationByAstNodeId = new Map<string, messages.ILocation>()
-  private readonly gherkinStepById = new Map<
+  private readonly gherkinStepByAstNodeId = new Map<
     string,
     messages.GherkinDocument.Feature.IStep
   >()
@@ -18,6 +19,11 @@ export default class Query {
 
   private readonly pickleStepIdsByAstNodeId = new Map<string, string[]>()
 
+  private readonly stepByUriLocation = new Map<
+    string,
+    messages.GherkinDocument.Feature.IStep
+  >()
+
   /**
    * Gets the location (line and column) of an AST node.
    * @param astNodeId
@@ -26,31 +32,46 @@ export default class Query {
     return this.locationByAstNodeId.get(astNodeId)
   }
 
-  public getGherkinDocuments(): ReadonlyArray<messages.IGherkinDocument> {
+  public getSources(): readonly messages.ISource[] {
+    return this.sources
+  }
+
+  public getGherkinDocuments(): readonly messages.IGherkinDocument[] {
     return this.gherkinDocuments
   }
 
-  public getPickles(): ReadonlyArray<messages.IPickle> {
+  public getPickles(): readonly messages.IPickle[] {
     return this.pickles
+  }
+
+  public getStep(
+    uri: string,
+    line: number
+  ): messages.GherkinDocument.Feature.IStep | undefined {
+    return this.stepByUriLocation.get([uri, line].join(':'))
   }
 
   /**
    * Gets all the pickle IDs
    * @param uri - the URI of the document
-   * @param lineNumber - optionally restrict results to a particular line number
+   * @param astNodeId - optionally restrict results to a particular AST node
    */
-  public getPickleIds(uri: string, astNodeId?: string): ReadonlyArray<string> {
+  public getPickleIds(uri: string, astNodeId?: string): readonly string[] {
     const pickleIdsByAstNodeId = this.pickleIdsMapByUri.get(uri)
     return astNodeId === undefined
       ? Array.from(new Set(pickleIdsByAstNodeId.values()))
       : pickleIdsByAstNodeId.get(astNodeId)
   }
 
-  public getPickleStepIds(astNodeId: string): ReadonlyArray<string> {
+  public getPickleStepIds(astNodeId: string): readonly string[] {
     return this.pickleStepIdsByAstNodeId.get(astNodeId) || []
   }
 
   public update(message: messages.IEnvelope): Query {
+    if (message.source) {
+      this.sources.push(message.source)
+    }
+
     if (message.gherkinDocument) {
       this.gherkinDocuments.push(message.gherkinDocument)
 
@@ -62,22 +83,34 @@ export default class Query {
 
         for (const featureChild of message.gherkinDocument.feature.children) {
           if (featureChild.background) {
-            this.updateGherkinBackground(featureChild.background)
+            this.updateGherkinBackground(
+              message.gherkinDocument.uri,
+              featureChild.background
+            )
           }
 
           if (featureChild.scenario) {
-            this.updateGherkinScenario(featureChild.scenario)
+            this.updateGherkinScenario(
+              message.gherkinDocument.uri,
+              featureChild.scenario
+            )
           }
 
           if (featureChild.rule) {
             const ruleChildren = featureChild.rule.children
             for (const ruleChild of ruleChildren) {
               if (ruleChild.background) {
-                this.updateGherkinBackground(ruleChild.background)
+                this.updateGherkinBackground(
+                  message.gherkinDocument.uri,
+                  ruleChild.background
+                )
               }
 
               if (ruleChild.scenario) {
-                this.updateGherkinScenario(ruleChild.scenario)
+                this.updateGherkinScenario(
+                  message.gherkinDocument.uri,
+                  ruleChild.scenario
+                )
               }
             }
           }
@@ -94,19 +127,21 @@ export default class Query {
   }
 
   private updateGherkinBackground(
+    uri: string,
     background: messages.GherkinDocument.Feature.IBackground
   ) {
     for (const step of background.steps) {
-      this.updateGherkinStep(step)
+      this.updateGherkinStep(uri, step)
     }
   }
 
   private updateGherkinScenario(
+    uri: string,
     scenario: messages.GherkinDocument.Feature.IScenario
   ) {
     this.locationByAstNodeId.set(scenario.id, scenario.location)
     for (const step of scenario.steps) {
-      this.updateGherkinStep(step)
+      this.updateGherkinStep(uri, step)
     }
 
     for (const examples of scenario.examples) {
@@ -116,9 +151,14 @@ export default class Query {
     }
   }
 
-  private updateGherkinStep(step: messages.GherkinDocument.Feature.IStep) {
+  private updateGherkinStep(
+    uri: string,
+    step: messages.GherkinDocument.Feature.IStep
+  ) {
     this.locationByAstNodeId.set(step.id, step.location)
-    this.gherkinStepById.set(step.id, step)
+    this.gherkinStepByAstNodeId.set(step.id, step)
+    const uriLocation = [uri, step.location.line].join(':')
+    this.stepByUriLocation.set(uriLocation, step)
   }
 
   private updatePickle(pickle: messages.IPickle) {