+++ react/javascript/src/components/gherkin/HookStep.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { messages } from '@cucumber/messages'
 import CucumberQueryContext from '../../CucumberQueryContext'
 import ErrorMessage from './ErrorMessage'
-import StepContainer from './StepContainer'
+import IconStep from './IconStep'
 import Attachment from './Attachment'
 
 interface IProps {
@@ -30,13 +30,13 @@ const HookStep: React.FunctionComponent<IProps> = ({ step }) => {
         hook.sourceReference.javaMethod.methodName
       : 'Unknown location'
     return (
-      <StepContainer status={stepResult.status}>
+      <IconStep status={stepResult.status}>
         <h3>Hook failed: {location}</h3>
         {stepResult.message && <ErrorMessage message={stepResult.message} />}
         {attachments.map((attachment, i) => (
           <Attachment key={i} attachment={attachment} />
         ))}
-      </StepContainer>
+      </IconStep>
     )
   }