+++ react/javascript/test/StepLineTest.tsx
@@ -3,15 +3,15 @@ import ReactDOM from 'react-dom'
 import React from 'react'
 import { messages } from '@cucumber/messages'
 import { Query as CucumberQuery } from '@cucumber/query'
-import Step from '../src/components/gherkin/Step'
 import { JSDOM } from 'jsdom'
 
 import UriContext from '../src/UriContext'
 import GherkinQueryContext from '../src/GherkinQueryContext'
 import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
 import CucumberQueryContext from '../src/CucumberQueryContext'
+import StepLine from '../src/components/gherkin/StepLine'
 
-describe('<Step>', () => {
+describe('<StepLine>', () => {
   it('renders', () => {
     const dom = new JSDOM(
       '<html lang="en"><body><div id="content"></div></body></html>'
@@ -59,11 +59,7 @@ describe('<Step>', () => {
       <GherkinQueryContext.Provider value={new StubGherkinQuery()}>
         <UriContext.Provider value={'some.feature'}>
           <CucumberQueryContext.Provider value={new StubCucumberQuery()}>
-            <Step
-              step={step}
-              renderStepMatchArguments={true}
-              renderMessage={true}
-            />
+            <StepLine step={step} renderStepMatchArguments={true} />
           </CucumberQueryContext.Provider>
         </UriContext.Provider>
       </GherkinQueryContext.Provider>
@@ -71,14 +67,12 @@ describe('<Step>', () => {
     ReactDOM.render(app, document.getElementById('content'))
 
     const plainTexts = Array.from(
-      document.querySelectorAll(
-        '#content h3 .cucumber-step__text, #content h3 .cucumber-step__keyword'
-      )
+      document.querySelectorAll('.cucumber-step__text, .cucumber-step__keyword')
     ).map((span) => span.textContent)
     assert.deepStrictEqual(plainTexts, ['Given', 'the ', ' pixies'])
 
     const paramTexts = Array.from(
-      document.querySelectorAll('#content h3 .cucumber-step__param')
+      document.querySelectorAll('.cucumber-step__param')
     ).map((a) => a.textContent)
     assert.deepStrictEqual(paramTexts, ['48'])
   })