+++ react/javascript/test/search/SearchTest.ts
@@ -1,7 +1,7 @@
 import assert from 'assert'
 
 import { IdGenerator } from '@cucumber/messages'
-import { generateMessages } from '@cucumber/gherkin'
+import { generateMessages, GherkinMediaType } from '@cucumber/gherkin'
 import Search from '../../src/search/Search'
 import { pretty, Query as GherkinQuery } from '@cucumber/gherkin-utils'
 
@@ -26,12 +26,17 @@ describe('Search', () => {
   })
 
   function prettyResults(feature: string, query: string): string {
-    const envelopes = generateMessages(feature, 'test.feature', {
-      includeGherkinDocument: true,
-      includePickles: true,
-      includeSource: true,
-      newId: IdGenerator.incrementing(),
-    })
+    const envelopes = generateMessages(
+      feature,
+      'test.feature',
+      GherkinMediaType.PLAIN,
+      {
+        includeGherkinDocument: true,
+        includePickles: true,
+        includeSource: true,
+        newId: IdGenerator.incrementing(),
+      }
+    )
     for (const envelope of envelopes) {
       gherkinQuery.update(envelope)
     }