+++ react/javascript/test/search/TagSearchTest.ts
@@ -1,6 +1,6 @@
 import { IdGenerator } from '@cucumber/messages'
 import assert from 'assert'
-import { generateMessages } from '@cucumber/gherkin'
+import { generateMessages, GherkinMediaType } from '@cucumber/gherkin'
 import { pretty, Query as GherkinQuery } from '@cucumber/gherkin-utils'
 import TagSearch from '../../src/search/TagSearch'
 
@@ -26,11 +26,16 @@ Feature: Solar System
   })
 
   function prettyResults(feature: string, query: string): string {
-    const envelopes = generateMessages(feature, 'test.feature', {
-      includePickles: true,
-      includeGherkinDocument: true,
-      newId: IdGenerator.incrementing(),
-    })
+    const envelopes = generateMessages(
+      feature,
+      'test.feature',
+      GherkinMediaType.PLAIN,
+      {
+        includePickles: true,
+        includeGherkinDocument: true,
+        newId: IdGenerator.incrementing(),
+      }
+    )
     for (const envelope of envelopes) {
       gherkinQuery.update(envelope)
     }